fix upload, except cancel
