+++ interface/package.json
@@ -27,13 +27,12 @@
     "@mui/material": "^5.13.6",
     "@table-library/react-table-library": "4.1.4",
     "@types/lodash-es": "^4.17.7",
-    "@types/node": "^20.3.2",
+    "@types/node": "^20.3.3",
     "@types/react": "^18.2.14",
     "@types/react-dom": "^18.2.6",
     "@types/react-router-dom": "^5.3.3",
-    "alova": "^2.8.0",
+    "alova": "^2.8.1",
     "async-validator": "^4.2.5",
-    "dev": "^0.1.3",
     "history": "^5.3.0",
     "jwt-decode": "^3.1.2",
     "lodash-es": "^4.17.21",
@@ -42,7 +41,7 @@
     "react-dom": "latest",
     "react-dropzone": "^14.2.3",
     "react-icons": "^4.10.1",
-    "react-router-dom": "^6.14.0",
+    "react-router-dom": "^6.14.1",
     "react-toastify": "^9.1.3",
     "sockette": "^2.0.6",
     "typesafe-i18n": "^5.24.3",
@@ -53,7 +52,7 @@
     "@typescript-eslint/eslint-plugin": "^5.60.1",
     "@typescript-eslint/parser": "^5.60.1",
     "@vitejs/plugin-react-swc": "^3.3.2",
-    "eslint": "^8.43.0",
+    "eslint": "^8.44.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-config-prettier": "^8.8.0",
