+++ interface/src/components/upload/index.ts
@@ -1,2 +1,1 @@
 export { default as SingleUpload } from './SingleUpload';
-export { default as useFileUpload } from './useFileUpload';
