+++ src/version.h
@@ -1,1 +1,1 @@
-#define EMSESP_APP_VERSION "3.6.0-dev.13a"
+#define EMSESP_APP_VERSION "3.6.0-dev.13b"
